Move ComposeFeatureFlags to base package
